refactor(convert): split the row builder out of the view
`view.rs` sat at 144 lines of the 150-line budget, and the two halves answer
different questions anyway: the view is what the window shows about the document,
the rows are what it shows about each file in a batch. They grow for different
reasons too — a field on the view, a column here.

`Row` and the builder move to `session/view/rows.rs`, the `foo.rs` + `foo/`
shape used elsewhere in the workspace. `view.rs` keeps `pub use rows::Row`, so
`lib.rs` and the three shells see no difference: 144 budgeted lines become 107
and 55.

No behaviour change; the existing session tests are untouched and still pass.
